+++ tests/test_mps.py
@@ -241,6 +241,14 @@ def test_canonical_form(bc):
     assert np.max(psi.norm_test()) < 1.e-14
 
 
+def test_increase_L():
+    s = site.SpinHalfSite(conserve='Sz')
+    psi = mps.MPS.from_product_state([s] * 3, ['up', 'down', 'up'], bc='infinite')
+    psi0 = psi.copy()
+    psi.increase_L(9)
+    expval = psi.expectation_value('Sigmaz')
+    npt.assert_equal(expval, [1., -1., 1.]*3)
+
 def test_group():
     s = site.SpinHalfSite(conserve='parity')
     psi1 = mps.MPS.from_singlets(s, 6, [(1, 3), (2, 5)], lonely=[0, 4], bc='finite')